- Fix the Pagination Issue - Refactoring the path, so it will be easy for laravel mix to run
